requirements.txt
beautifulsoup4==4.13.4
flask-cors==5.0.1
fuzzywuzzy==0.18.0
numpy==2.2.5
openai==1.78.1
pdfplumber==0.11.6
python-dotenv==1.1.0
readability-lxml==0.8.4.1
requests==2.32.3
tiktoken==0.9.0
